wasm-mutate: Add a mutator to modify data segments (#447)
* wasm-mutate: Add a mutator to modify data segments

In an effort to minimize modules this commit adds a mutator which will
modify data segments to either add or remove bytes (although only
removal is used when reducing). The main meat of this commit is
refactoring out the translation from `wasmparser` types to `wasm-encoder`
types into a shared module for both the `RemoveItem` mutator and this
new mutator to use.

* Use `raw_mutate_func` for data mutations

* Fix tests

Also make the rename export mutator a bit more robust

* Factor in a max size for data mutations
